Add product as a new property for main component
By adding ‘product’ as a property, state changes in the caller (user of NumPad) can be reflected in the NumPad. Although product is now passed as a prop, it still needs to be saved in NumPad local state. The reason is that the scanning function compares old and new product to decide if quantity will be increased.